adding blocked preview
