+++ feature/settings/src/main/java/org/mozilla/social/feature/settings/contentpreferences/blockedusers/BlockedUsersSettingsScreen.kt
@@ -1,25 +1,25 @@
 package org.mozilla.social.feature.settings.contentpreferences.blockedusers
 
-import androidx.annotation.StringRes
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.tooling.preview.Preview
 import androidx.paging.PagingData
 import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.emptyFlow
 import org.koin.androidx.compose.koinViewModel
-import org.mozilla.social.common.utils.StringFactory
+import org.mozilla.social.core.navigation.navigationModule
 import org.mozilla.social.core.ui.common.MoSoSurface
 import org.mozilla.social.core.ui.common.account.toggleablelist.ToggleableAccountList
 import org.mozilla.social.core.ui.common.account.toggleablelist.ToggleableAccountListItemState
-import org.mozilla.social.core.ui.common.account.toggleablelist.ToggleableButtonState
-import org.mozilla.social.core.ui.common.button.MoSoButtonPrimaryDefaults
-import org.mozilla.social.core.ui.common.button.MoSoButtonSecondaryDefaults
-import org.mozilla.social.core.ui.common.button.MoSoButtonTheme
+import org.mozilla.social.core.ui.common.utils.PreviewTheme
 import org.mozilla.social.feature.settings.R
 import org.mozilla.social.feature.settings.ui.SettingsColumn
 
 @Composable
-internal fun BlockedUsersSettingsScreen(viewModel: BlockedUsersViewModel = koinViewModel()) {
+internal fun BlockedUsersSettingsScreen(
+    viewModel: BlockedUsersViewModel = koinViewModel()
+) {
     BlockedUsersSettingsScreen(
         pagingData = viewModel.blocks,
         blockedUsersInteractions = viewModel,
@@ -46,24 +46,19 @@ private fun BlockedUsersSettingsScreen(
     }
 }
 
-/**
- * Blocked button state is an implementation of [ToggleableButtonState] which encapsulates the
- * different states a user can be in, and what the text on the button should be for those states
- */
-sealed class BlockedButtonState(
-    @get:StringRes override val text: Int,
-    override val confirmationText: StringFactory? = null,
-    override val theme: MoSoButtonTheme,
-) : ToggleableButtonState {
-    data object Blocked : BlockedButtonState(
-        text = R.string.unblock,
-        theme = MoSoButtonSecondaryDefaults,
+@Preview
+@Composable
+private fun BlockedUsersSettingsScreenPreview() {
+    PreviewTheme(
+        modules = listOf(navigationModule)
     ) {
-        override val confirmationText: StringFactory? = null
+        BlockedUsersSettingsScreen(
+            pagingData = emptyFlow(),
+            blockedUsersInteractions = object : BlockedUsersInteractions {
+                override fun onScreenViewed() = Unit
+                override fun onButtonClicked(accountId: String, buttonState: BlockedButtonState) = Unit
+                override fun onAccountClicked(accountId: String) = Unit
+            }
+        )
     }
-
-    data class Unblocked(override val confirmationText: StringFactory?) : BlockedButtonState(
-        text = R.string.block,
-        theme = MoSoButtonPrimaryDefaults,
-    )
 }
