+++ feature/settings/src/main/java/org/mozilla/social/feature/settings/contentpreferences/mutedusers/MutedUsersInteractions.kt
@@ -2,4 +2,6 @@ package org.mozilla.social.feature.settings.contentpreferences.mutedusers
 
 interface MutedUsersInteractions {
     fun onScreenViewed()
+    fun onButtonClicked(accountId: String, mutedButtonState: MutedButtonState)
+    fun onAccountClicked(accountId: String)
 }
